backend/cloud-functions/get-user-data-function/requirements.txt
--index-url https://pypi.org/simple/

# For Firebase/Firestore
google-cloud-firestore

# For Gemini API (Vertex AI)
google-cloud-aiplatform

# General GCP Functions Framework
functions-framework